Remove cli related code
