+++ pom.xml
@@ -5,7 +5,7 @@
 
   <groupId>io.github.metal_pony</groupId>
   <artifactId>sudoku</artifactId>
-  <version>1.0.0</version>
+  <version>1.1.0-SNAPSHOT</version>
 
   <name>sudoku</name>
