--- scripts/build
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-# Builds an executable jar of Sudoku utilities.
-# Run this from the base project directory.
-
-./mvnw package -f pom-exe.xml "$@"
