--- scripts/build-graal
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-# Builds a native executable for the current system with GraalVM.
-# Ensure that the appropriate GraalVM jdk is installed and the default.
-
-./mvnw clean package -f pom-exe-graal.xml "$@"
