--- scripts/play
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-# Starts a Sudoku game in a Java Swing GUI.
-# Run this from the base project directory.
-
-java -jar target/sudoku-1.jar play "$@"
