--- scripts/rebuildAndRun
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-# Generates Sudoku configurations.
-# Run this from the base project directory.
-
-./mvnw package -f pom-exe.xml && java -jar target/sudoku-1.jar "$@"
