--- scripts/solve
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-# Solves a given Sudoku puzzle.
-# Run this from the base project directory.
-
-java -jar target/sudoku-1.jar solve "$@"
